Tweak message and temporarily set model ttl to 0

+++ components/resource/resourcemanager.hpp
@@ -60,7 +60,7 @@ namespace Resource
 
         /// How long to keep objects in cache after no longer being referenced.
         void setExpiryDelay(double expiryDelay) override { mExpiryDelay = expiryDelay; }
-        float getExpiryDelay() const { return mExpiryDelay; }
+        double getExpiryDelay() const { return mExpiryDelay; }
 
         const VFS::Manager* getVFS() const { return mVFS; }
